requirement.txt
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib
python-dotenv
flask
google-genai

